Move MultiFetchRequest and TimeFrame to dedicated file
