+++ render/data/query.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"net/http"
 	"strings"
 	"sync"
 	"time"
@@ -33,29 +32,6 @@ FORMAT RowBinary`
 
 // table, prewhere, where
 const queryUnaggregated = `SELECT Path, groupArray(Time), groupArray(Value), groupArray(Timestamp) FROM %s %s %s GROUP BY Path FORMAT RowBinary`
-
-// TimeFrame contains information about fetch request time conditions
-type TimeFrame struct {
-	From          int64
-	Until         int64
-	MaxDataPoints int64
-}
-
-// MultiFetchRequest is a map of TimeFrame keys and targets slice of strings values
-type MultiFetchRequest map[TimeFrame]*Targets
-
-func (m *MultiFetchRequest) checkMetricsLimitExceeded(num int) error {
-	if num <= 0 {
-		// zero or negative means unlimited
-		return nil
-	}
-	for _, t := range *m {
-		if num < t.AM.Len() {
-			return clickhouse.NewErrorWithCode(fmt.Sprintf("metrics limit exceeded: %d < %d", num, t.AM.Len()), http.StatusForbidden)
-		}
-	}
-	return nil
-}
 
 // CHResponse contains the parsed Data and From/Until timestamps
 type CHResponse struct {
@@ -106,57 +82,6 @@ func (q *query) appendReply(chr CHResponse) {
 	q.lock.Lock()
 	q.CHResponses = append(q.CHResponses, chr)
 	q.lock.Unlock()
-}
-
-// Fetch fetches the parsed ClickHouse data returns CHResponses
-func (m *MultiFetchRequest) Fetch(ctx context.Context, cfg *config.Config, chContext string) (CHResponses, error) {
-	var lock sync.RWMutex
-	var wg sync.WaitGroup
-	logger := scope.Logger(ctx)
-	setCarbonlinkClient(&cfg.Carbonlink)
-
-	err := m.checkMetricsLimitExceeded(cfg.Common.MaxMetricsPerTarget)
-	if err != nil {
-		logger.Error("data fetch", zap.Error(err))
-		return nil, err
-	}
-
-	ctxTimeout, cancel := context.WithTimeout(ctx, cfg.ClickHouse.DataTimeout.Duration)
-	defer cancel()
-
-	errors := make([]error, 0, len(*m))
-	query := newQuery(cfg, len(*m))
-
-	for tf, targets := range *m {
-		if tf.MaxDataPoints <= 0 || int64(cfg.ClickHouse.MaxDataPoints) < tf.MaxDataPoints {
-			tf.MaxDataPoints = int64(cfg.ClickHouse.MaxDataPoints)
-		}
-		err := targets.selectDataTable(cfg, tf.From, tf.Until, chContext)
-		if err != nil {
-			lock.Lock()
-			errors = append(errors, err)
-			lock.Unlock()
-			logger.Error("data tables is not specified", zap.Error(err))
-			return EmptyResponse(), err
-		}
-		wg.Add(1)
-		go func(tf TimeFrame, targets *Targets) {
-			defer wg.Done()
-			err := query.getDataPoints(ctxTimeout, tf, targets)
-			if err != nil {
-				lock.Lock()
-				errors = append(errors, err)
-				lock.Unlock()
-				return
-			}
-		}(tf, targets)
-	}
-	wg.Wait()
-	for len(errors) != 0 {
-		return EmptyResponse(), errors[0]
-	}
-
-	return query.CHResponses, nil
 }
 
 func (q *query) getDataPoints(ctx context.Context, tf TimeFrame, targets *Targets) error {
